test(daemon): split the three over-budget daemon test files along their source modules
snapshot-handler.test.ts, session-replay-repair-transaction.test.ts, and
find.test.ts sat over the 1,000-line test-file-size-ratchet tripwire with no
headroom (#2346 follow-up). Each mirrors a single source module through an
integration-style suite, so each splits by concern rather than by a second
source file:

- snapshot-handler.test.ts keeps core snapshot/diff routing and policy;
  snapshot-handler-freshness.test.ts takes annotations, Android timeout
  evidence, and staleness/visibility warnings; snapshot-handler-capture-retry
  takes captureSnapshot's pending-outcome and stabilization retries;
  snapshot-handler-wait.test.ts takes the wait command. The shared handler
  wrapper moved into the existing snapshot-handler.fixtures.ts sibling.
- find.test.ts keeps core click/focus/type dispatch; find-ambiguous-match,
  find-read-only, and find-production-route take the AMBIGUOUS_MATCH/list
  behavior, read-only find (wait/exists), and the #1654 real-interaction-leaf
  route respectively. The shared click-scenario runner moved into the
  existing find-touch-runtime-fixture.ts sibling.
- session-replay-repair-transaction.test.ts keeps the core arm/resume/reap
  lifecycle; -force takes the #1258 --force/--overwrite preflight; and
  -close-ordering takes the BLOCKER 2/3 platform-close sequencing. The
  shared setup/handleCloseCommand/makeCompleteRepairSession helpers moved
  into the existing session-replay-repair.fixtures.ts sibling.

Every moved test keeps its name and body unchanged; verified by diffing each
test's body against the pre-split file. No assertion changed and no test was
dropped, merged, or added.

Two other daemon test files (session-replay-divergence.test.ts,
session-replay-runtime-maestro.test.ts) are also over the ratchet tripwire,
but were already over it before #2346 and are unrelated to this follow-up —
left for a separate issue rather than folded in here.

Closes #2353
