+++ src/daemon/handlers/__tests__/snapshot-handler.test.ts
@@ -1,40 +1,16 @@
 import { test, expect, vi, afterEach, beforeEach } from 'vitest';
 import { legacyDispatchCapture } from '../../__tests__/legacy-snapshot-capture-fixture.ts';
-import {
-  getRuntimeBindings,
-  mockTapPoint,
-  resetGetRuntimeFixture,
-} from '../../__tests__/interaction-get-runtime-fixture.ts';
-import fs from 'node:fs';
-import { handleSnapshotCommands as handleProductionSnapshotCommands } from '../snapshot.ts';
-import { captureSnapshot } from '../../snapshot-capture.ts';
-import { SessionStore } from '../../session-store.ts';
+import { resetGetRuntimeFixture } from '../../__tests__/interaction-get-runtime-fixture.ts';
+import { resetSnapshotRuntimeFixture } from '../../__tests__/snapshot-runtime-fixture.ts';
 import { setActiveProviderDeviceRuntimes } from '../../../provider-device-runtime.ts';
 import type { DaemonResponse } from '../../daemon-request.ts';
 import type { SessionState } from '../../session-state.ts';
-import { AppError } from '@agent-device/kernel/errors';
-import { platformResourceCleanup } from '../../../platform-runtime-resource-cleanup.ts';
-import { buildInteractionSurfaceSignature } from '../../interaction-outcome-policy.ts';
-import { buildSnapshotPresentationKey } from '@agent-device/kernel/snapshot';
-import { snapshotCliOutput } from '../../../commands/capture/output.ts';
-import type { CaptureSnapshotResult } from '@agent-device/contracts/client';
-import { buildNodes } from '../../../__tests__/test-utils/snapshot-builders.ts';
-import {
-  fixtureScreenshotCaptures,
-  resetSnapshotRuntimeFixture,
-  snapshotRuntimeFixture,
-} from '../../__tests__/snapshot-runtime-fixture.ts';
 import {
   androidCapture,
   androidDevice,
-  androidTextRows,
-  batteryCapture,
   countingSnapshotRuntime,
-  inboxBaselineNodes,
-  inboxRow,
+  handleSnapshotCommands,
   iosSimulatorDevice,
-  locationRequiredCapture,
-  makeAndroidFreshnessSession,
   makeProviderRuntimeOwning,
   makeSession,
   makeSessionStore,
@@ -67,31 +43,6 @@ import { expireRefFrame, refFrame, refFrameState, refFrameTree } from '../../ref
 const mockRunnerCommand = vi.mocked(runAppleRunnerCommand);
 const mockBuildIosOpenCommandHint = vi.mocked(buildIosOpenCommandHint);
 
-const SNAPSHOT_ROUTE_RUNTIME_COMMANDS = new Set(['snapshot', 'diff', 'settings', 'alert']);
-
-function handleSnapshotCommands(
-  params: Parameters<typeof handleProductionSnapshotCommands>[0],
-): ReturnType<typeof handleProductionSnapshotCommands> {
-  if (!SNAPSHOT_ROUTE_RUNTIME_COMMANDS.has(params.req.command)) {
-    return handleProductionSnapshotCommands({
-      ...params,
-      platformResourceCleanup: params.platformResourceCleanup ?? platformResourceCleanup,
-    });
-  }
-  const runtime = snapshotRuntimeFixture(params.req.meta?.requestId);
-  return handleProductionSnapshotCommands({
-    ...params,
-    inspectFacts: params.inspectFacts ?? runtime.inspectFacts,
-    bindDevice: params.bindDevice ?? runtime.bindDevice,
-    platformResourceCleanup: params.platformResourceCleanup ?? platformResourceCleanup,
-  });
-}
-
-// An Apple wait runs inside an opened app: that bundle id is XCUITest's attach target, and
-// without one the plan asks for the without-active-app row local Apple refuses.
-const appAttach = (d: SessionState['device']): Partial<SessionState> =>
-  d.platform === 'apple' ? { appBundleId: 'com.example.app' } : {};
-
 afterEach(() => {
   setActiveProviderDeviceRuntimes([]);
 });
@@ -106,166 +57,6 @@ beforeEach(() => {
   mockBuildIosOpenCommandHint.mockReset();
   mockBuildIosOpenCommandHint.mockResolvedValue(undefined);
 });
-
-function makeAndroidTimeoutEvidenceSession(sessionName: string): SessionStore {
-  const sessionStore = makeSessionStore();
-  const session = makeSession(sessionName, androidDevice);
-  session.snapshot = {
-    nodes: [
-      {
-        ref: 'e1',
-        index: 0,
-        depth: 0,
-        type: 'android.widget.Button',
-        label: 'Continue',
-        hittable: true,
-        rect: { x: 20, y: 40, width: 120, height: 48 },
-      },
-    ],
-    createdAt: Date.now(),
-    backend: 'android',
-  };
-  sessionStore.set(sessionName, session);
-  return sessionStore;
-}
-
-function mockAndroidTimeoutEvidenceDispatch(): void {
-  legacyDispatchCapture.mockImplementation(async (_device, command) => {
-    if (command === 'snapshot') throw androidSnapshotTimeoutError();
-    return {};
-  });
-}
-
-function androidSnapshotTimeoutError(): AppError {
-  return new AppError(
-    'COMMAND_FAILED',
-    'Android snapshot helper timed out while waiting for the UI to become idle.',
-    {
-      cmd: 'adb',
-      args: ['shell', 'am', 'instrument'],
-      androidCaptureFailureReason: 'accessibility-timeout',
-      hint: 'Android accessibility snapshots can be blocked by busy or continuously changing app UI. Use screenshot as visual truth after this timeout.',
-    },
-  );
-}
-
-function expectAndroidTimeoutEvidence(
-  response: Awaited<ReturnType<typeof handleSnapshotCommands>>,
-) {
-  if (!response) throw new Error('Expected snapshot response');
-  if (response.ok) throw new Error('Expected snapshot timeout failure');
-  expect(response.error.message).toMatch(/snapshot helper timed out/i);
-  expect(response.error.hint).toMatch(/Use screenshot as visual truth/i);
-  assertAndroidTimeoutEvidencePayload(response.error.details?.androidSnapshotTimeoutScreenshot);
-}
-
-function assertAndroidTimeoutEvidencePayload(evidence: unknown) {
-  if (!evidence || typeof evidence !== 'object') {
-    throw new Error('Expected Android snapshot timeout screenshot evidence');
-  }
-  const record = evidence as Record<string, unknown>;
-  expect(record.path).toEqual(expect.stringContaining('snapshot-timeout-overlay-refs.png'));
-  expect(fs.existsSync(record.path as string)).toBe(true);
-  expect(record.overlayRefsAnnotated).toBe(true);
-  expect(record.overlayRefs).toEqual([expect.objectContaining({ ref: 'e1', label: 'Continue' })]);
-}
-
-async function runWaitCommand(
-  sessionName: string,
-  device: SessionState['device'],
-  positionals: string[],
-) {
-  const sessionStore = makeSessionStore();
-  sessionStore.set(sessionName, makeSession(sessionName, device, appAttach(device)));
-  return await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'wait', { positionals }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-    ...snapshotRuntimeFixture(),
-  });
-}
-
-const locationPermissionNodes = [
-  {
-    index: 0,
-    depth: 0,
-    type: 'android.widget.FrameLayout',
-    label: 'Location permission',
-    rect: { x: 0, y: 0, width: 390, height: 844 },
-  },
-  {
-    index: 1,
-    depth: 1,
-    parentIndex: 0,
-    type: 'android.widget.TextView',
-    label: 'Allow location access?',
-    rect: { x: 24, y: 210, width: 342, height: 40 },
-  },
-  {
-    index: 2,
-    depth: 1,
-    parentIndex: 0,
-    type: 'android.widget.Button',
-    label: 'Not now',
-    rect: { x: 24, y: 320, width: 140, height: 48 },
-    hittable: true,
-  },
-  {
-    index: 3,
-    depth: 1,
-    parentIndex: 0,
-    type: 'android.widget.Button',
-    label: 'Continue',
-    rect: { x: 180, y: 320, width: 160, height: 48 },
-    hittable: true,
-  },
-];
-
-const iosSurfaceSummaryNodes = [
-  {
-    index: 0,
-    depth: 0,
-    type: 'XCUIElementTypeApplication',
-    label: 'Expo Go',
-    rect: { x: 0, y: 0, width: 393, height: 852 },
-  },
-  {
-    index: 1,
-    depth: 1,
-    type: 'XCUIElementTypeImage',
-    label: 'gearshape.fill',
-    rect: { x: 12, y: 54, width: 24, height: 24 },
-  },
-  {
-    index: 2,
-    depth: 1,
-    type: 'XCUIElementTypeOther',
-    label: 'Tab Bar',
-    rect: { x: 0, y: 760, width: 393, height: 92 },
-  },
-  {
-    index: 3,
-    depth: 1,
-    type: 'XCUIElementTypeStaticText',
-    label: 'Confirm catalog refresh',
-    rect: { x: 48, y: 280, width: 297, height: 36 },
-  },
-  {
-    index: 4,
-    depth: 1,
-    type: 'XCUIElementTypeButton',
-    label: 'Keep browsing',
-    rect: { x: 48, y: 360, width: 297, height: 48 },
-  },
-  {
-    index: 5,
-    depth: 1,
-    type: 'XCUIElementTypeButton',
-    identifier: 'host.exp.exponent:id/reload_button',
-    rect: { x: 260, y: 54, width: 48, height: 48 },
-  },
-];
 
 test('snapshot rejects @ref scope without existing session snapshot', async () => {
   const sessionStore = makeSessionStore();
@@ -607,838 +398,6 @@ test('snapshot surfaces filtered-to-zero Android guidance for interactive snapsh
   }
 });
 
-test('snapshot annotations survive pending interaction capture into CLI JSON', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'android-interaction-annotation-bundle';
-  const session = makeSession(sessionName, androidDevice);
-  const baselineNodes = [
-    {
-      ref: 'e1',
-      index: 0,
-      depth: 0,
-      type: 'android.widget.Button',
-      label: 'Open albums',
-      hittable: true,
-      rect: { x: 20, y: 120, width: 160, height: 48 },
-    },
-  ];
-  const changedNodes = [
-    {
-      index: 0,
-      depth: 0,
-      type: 'android.widget.TextView',
-      label: 'Albums',
-      rect: { x: 32, y: 240, width: 180, height: 52 },
-    },
-  ];
-  const snapshotQuality = { state: 'healthy', backend: 'tree' };
-  session.pendingInteractionOutcome = {
-    action: 'click',
-    command: 'press',
-    positionals: ['100', '144'],
-    flags: { platform: 'android' },
-    markedAt: Date.now(),
-    attemptsRemaining: 2,
-    preSignature: buildInteractionSurfaceSignature(baselineNodes),
-  };
-  sessionStore.set(sessionName, session);
-
-  legacyDispatchCapture.mockResolvedValue({
-    nodes: changedNodes,
-    truncated: false,
-    backend: 'android',
-    analysis: { rawNodeCount: 1, maxDepth: 0 },
-    quality: snapshotQuality,
-    warnings: ['backend warning from interaction capture'],
-  });
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'snapshot'),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(true);
-  if (!response?.ok) return;
-
-  expect(response.data?.snapshotQuality).toEqual(snapshotQuality);
-  expect(response.data?.warnings).toEqual(['backend warning from interaction capture']);
-
-  const cliOutput = await snapshotCliOutput({
-    result: response.data as unknown as CaptureSnapshotResult,
-  });
-  expect(cliOutput.jsonData).toMatchObject({
-    nodes: [expect.objectContaining({ label: 'Albums' })],
-    truncated: false,
-    snapshotQuality,
-    warnings: ['backend warning from interaction capture'],
-  });
-  expect(cliOutput.jsonData).not.toHaveProperty('analysis');
-  expect(cliOutput.jsonData).not.toHaveProperty('freshness');
-});
-
-test('snapshot timeout captures Android screenshot evidence with overlay refs', async () => {
-  const sessionName = 'android-timeout-evidence';
-  const sessionStore = makeAndroidTimeoutEvidenceSession(sessionName);
-  mockAndroidTimeoutEvidenceDispatch();
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'snapshot'),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-  expectAndroidTimeoutEvidence(response);
-  expect(legacyDispatchCapture.mock.calls.map((call) => call[1])).toEqual(['snapshot']);
-  expect(fixtureScreenshotCaptures.at(-1)?.options).toMatchObject({ stabilize: false });
-});
-
-test('snapshot warns when recent snapshot node count collapses sharply', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'android-stale-collapse';
-  const session = makeSession(sessionName, androidDevice);
-  session.snapshot = {
-    nodes: buildNodes(androidTextRows(50, (row) => `Row ${row}`)),
-    createdAt: Date.now(),
-    backend: 'android',
-  };
-  sessionStore.set(sessionName, session);
-
-  legacyDispatchCapture.mockResolvedValue(
-    androidCapture(
-      androidTextRows(8, (row) => `Next ${row}`),
-      { rawNodeCount: 8, maxDepth: 1 },
-    ),
-  );
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'snapshot'),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(true);
-  if (response?.ok) {
-    expect(response.data?.warnings).toEqual([
-      expect.stringContaining('Recent snapshots dropped sharply in node count'),
-    ]);
-  }
-});
-
-test('snapshot does not warn on expected node drop across presentation modes', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'ios-presentation-drop';
-  const session = makeSession(sessionName, iosSimulatorDevice);
-  session.appBundleId = 'com.example.app';
-  session.snapshot = {
-    nodes: Array.from({ length: 50 }, (_, index) => ({
-      ref: `e${index + 1}`,
-      index,
-      depth: 0,
-      type: 'StaticText',
-      label: `Row ${index + 1}`,
-    })),
-    createdAt: Date.now(),
-    backend: 'xctest',
-    presentationKey: buildSnapshotPresentationKey({ interactiveOnly: false }),
-  };
-  sessionStore.set(sessionName, session);
-
-  legacyDispatchCapture.mockResolvedValue({
-    nodes: Array.from({ length: 8 }, (_, index) => ({
-      index,
-      depth: 0,
-      type: 'Button',
-      label: `Action ${index + 1}`,
-      hittable: true,
-    })),
-    truncated: false,
-    backend: 'xctest',
-  });
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'snapshot', { flags: { snapshotInteractiveOnly: true } }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(true);
-  if (response?.ok) {
-    expect(response.data?.warnings ?? []).toEqual(
-      expect.not.arrayContaining([
-        expect.stringContaining('Recent snapshots dropped sharply in node count'),
-      ]),
-    );
-  }
-});
-
-test('snapshot automatically retries stale Android trees after recent navigation', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'android-stale-retries-to-fresh';
-  const session = makeAndroidFreshnessSession(sessionName, 'press', inboxBaselineNodes(24));
-  sessionStore.set(sessionName, session);
-
-  legacyDispatchCapture
-    .mockResolvedValueOnce(
-      androidCapture(androidTextRows(24, inboxRow), { rawNodeCount: 24, maxDepth: 2 }),
-    )
-    .mockResolvedValueOnce(
-      androidCapture(
-        [
-          { index: 0, depth: 0, type: 'android.widget.TextView', label: 'Create document' },
-          { index: 1, depth: 0, type: 'android.widget.Button', label: 'Submit', hittable: true },
-        ],
-        { rawNodeCount: 2, maxDepth: 1 },
-      ),
-    );
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'snapshot', { flags: { snapshotInteractiveOnly: true } }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(true);
-  if (response?.ok) {
-    expect(response.data?.warnings).toBeUndefined();
-    expect(response.data?.nodes).toEqual(
-      expect.arrayContaining([expect.objectContaining({ label: 'Create document' })]),
-    );
-  }
-  expect(legacyDispatchCapture).toHaveBeenCalledTimes(2);
-  expect(sessionStore.get(sessionName)?.androidSnapshotFreshness).toBeUndefined();
-});
-
-test('snapshot warns when Android freshness retries still return the previous route', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'android-stale-after-press';
-  const session = makeAndroidFreshnessSession(sessionName, 'press', inboxBaselineNodes(24));
-  sessionStore.set(sessionName, session);
-
-  legacyDispatchCapture.mockResolvedValue(
-    androidCapture(androidTextRows(24, inboxRow), { rawNodeCount: 24, maxDepth: 2 }),
-  );
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'snapshot', { flags: { snapshotInteractiveOnly: true } }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(true);
-  if (response?.ok) {
-    expect(response.data?.warnings).toEqual([
-      expect.stringContaining(
-        'Recent press was followed by a nearly identical snapshot after 3 automatic retries',
-      ),
-    ]);
-  }
-  expect(legacyDispatchCapture).toHaveBeenCalledTimes(4);
-});
-
-test('snapshot response includes normalized visibility metadata', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'android-visibility';
-  sessionStore.set(sessionName, makeSession(sessionName, androidDevice));
-
-  legacyDispatchCapture.mockResolvedValue({
-    nodes: [
-      {
-        index: 0,
-        depth: 0,
-        type: 'android.widget.ScrollView',
-        label: 'Messages',
-        rect: { x: 0, y: 100, width: 390, height: 500 },
-        hiddenContentBelow: true,
-      },
-      {
-        index: 1,
-        depth: 1,
-        parentIndex: 0,
-        type: 'android.widget.Button',
-        label: 'Visible message',
-        rect: { x: 0, y: 140, width: 390, height: 48 },
-        hittable: true,
-      },
-    ],
-    truncated: false,
-    backend: 'android',
-    analysis: { rawNodeCount: 2, maxDepth: 1 },
-  });
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'snapshot', { flags: { snapshotInteractiveOnly: true } }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(true);
-  if (response?.ok) {
-    expect(response.data?.visibility).toEqual({
-      partial: true,
-      visibleNodeCount: 2,
-      totalNodeCount: 2,
-      reasons: ['scroll-hidden-below'],
-    });
-  }
-});
-
-test('diff snapshot carries stale-tree warnings for recent Android presses', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'android-diff-stale-after-press';
-  const session = makeAndroidFreshnessSession(sessionName, 'press', inboxBaselineNodes(24));
-  sessionStore.set(sessionName, session);
-
-  legacyDispatchCapture.mockResolvedValue(
-    androidCapture(androidTextRows(24, inboxRow), { rawNodeCount: 24, maxDepth: 2 }),
-  );
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'diff', {
-      positionals: ['snapshot'],
-      flags: { snapshotInteractiveOnly: true },
-    }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(true);
-  if (response?.ok) {
-    expect(response.data?.warnings).toEqual([
-      expect.stringContaining(
-        'Recent press was followed by a nearly identical snapshot after 3 automatic retries',
-      ),
-    ]);
-  }
-  expect(legacyDispatchCapture).toHaveBeenCalledTimes(4);
-});
-
-test('Android ref refresh mode does not retry narrow snapshots as sharp drops', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'android-ref-refresh-no-sharp-drop';
-  const session = makeAndroidFreshnessSession(
-    sessionName,
-    'press',
-    buildNodes(androidTextRows(50, (row) => `Previous row ${row}`)),
-  );
-  sessionStore.set(sessionName, session);
-
-  legacyDispatchCapture.mockResolvedValue(
-    androidCapture(
-      Array.from({ length: 8 }, (_, index) => ({
-        index,
-        depth: 0,
-        type: 'android.widget.TextView',
-      })),
-      { rawNodeCount: 8, maxDepth: 1 },
-    ),
-  );
-
-  const result = await captureSnapshot({
-    device: androidDevice,
-    session,
-    flags: { snapshotInteractiveOnly: true },
-    logPath: '/tmp/daemon.log',
-    androidFreshnessMode: 'ref-refresh',
-  });
-
-  expect(result.freshness).toBeUndefined();
-  expect(legacyDispatchCapture).toHaveBeenCalledTimes(1);
-  expect(session.androidSnapshotFreshness).toBeUndefined();
-});
-
-test('captureSnapshot lazily retries pending no-change touch before returning fresh state', async () => {
-  const sessionName = 'ios-lazy-outcome-retry';
-  const session = makeSession(sessionName, iosSimulatorDevice);
-  const baselineNodes = [
-    {
-      ref: 'e1',
-      index: 0,
-      depth: 0,
-      type: 'Button',
-      label: 'Open feed',
-      identifier: 'open-feed',
-      hittable: true,
-      rect: { x: 20, y: 120, width: 160, height: 48 },
-    },
-  ];
-  session.snapshot = {
-    nodes: baselineNodes,
-    createdAt: Date.now(),
-    backend: 'xctest',
-  };
-  session.pendingInteractionOutcome = {
-    action: 'click',
-    command: 'press',
-    positionals: ['100', '144'],
-    flags: { platform: 'ios' },
-    markedAt: Date.now(),
-    attemptsRemaining: 2,
-    preSignature: [
-      {
-        key: 'open-feed|Open feed||Button||enabled|unselected|hittable|#0',
-        x: 20,
-        y: 120,
-        width: 160,
-        height: 48,
-        discriminating: true,
-      },
-    ],
-  };
-
-  let pressed = false;
-  mockTapPoint.mockImplementation(async () => {
-    pressed = true;
-    return { clicked: true };
-  });
-  legacyDispatchCapture.mockImplementation(async () => {
-    return {
-      nodes: !pressed
-        ? baselineNodes
-        : [
-            {
-              index: 0,
-              depth: 0,
-              type: 'Button',
-              label: 'Back',
-              identifier: 'back',
-              hittable: true,
-              rect: { x: 20, y: 60, width: 90, height: 44 },
-            },
-            {
-              index: 1,
-              depth: 0,
-              type: 'StaticText',
-              label: 'Feed',
-              rect: { x: 20, y: 140, width: 160, height: 48 },
-            },
-          ],
-      backend: 'xctest',
-    };
-  });
-
-  const result = await captureSnapshot({
-    device: iosSimulatorDevice,
-    session,
-    flags: { snapshotInteractiveOnly: true },
-    logPath: '/tmp/daemon.log',
-    ...getRuntimeBindings(),
-  });
-
-  expect(result.snapshot.nodes).toEqual(
-    expect.arrayContaining([expect.objectContaining({ label: 'Feed' })]),
-  );
-  // R58: the retry re-fires through the bound `tapPoint`, on the recorded coordinate pair.
-  expect(mockTapPoint).toHaveBeenCalledTimes(1);
-  expect(mockTapPoint.mock.calls[0]?.[0]?.point).toEqual({ x: 100, y: 144 });
-  expect(session.pendingInteractionOutcome).toBeUndefined();
-});
-
-test('captureSnapshot does not retry when a tap change appears after a short delay', async () => {
-  const sessionName = 'android-delayed-outcome-without-retry';
-  const session = makeSession(sessionName, androidDevice);
-  const baselineNodes = [
-    {
-      ref: 'e1',
-      index: 0,
-      depth: 0,
-      type: 'android.widget.Button',
-      label: 'Open drawer',
-      hittable: true,
-      rect: { x: 20, y: 120, width: 160, height: 48 },
-    },
-  ];
-  const changedNodes = [
-    {
-      index: 0,
-      depth: 0,
-      type: 'android.widget.TextView',
-      label: 'Albums',
-      rect: { x: 32, y: 240, width: 180, height: 52 },
-    },
-  ];
-  session.pendingInteractionOutcome = {
-    action: 'click',
-    command: 'press',
-    positionals: ['100', '144'],
-    flags: { platform: 'android' },
-    markedAt: Date.now(),
-    attemptsRemaining: 2,
-    preSignature: buildInteractionSurfaceSignature(baselineNodes),
-  };
-
-  let snapshotCalls = 0;
-  legacyDispatchCapture.mockImplementation(async (_device, command) => {
-    expect(command).toBe('snapshot');
-    snapshotCalls += 1;
-    return {
-      nodes: snapshotCalls === 1 ? baselineNodes : changedNodes,
-      backend: 'android',
-    };
-  });
-
-  const result = await captureSnapshot({
-    device: androidDevice,
-    session,
-    flags: { snapshotInteractiveOnly: true },
-    logPath: '/tmp/daemon.log',
-  });
-
-  expect(result.snapshot.nodes).toEqual(
-    expect.arrayContaining([expect.objectContaining({ label: 'Albums' })]),
-  );
-  expect(legacyDispatchCapture.mock.calls.map((call) => call[1])).toEqual(['snapshot', 'snapshot']);
-  expect(session.pendingInteractionOutcome).toBeUndefined();
-});
-
-test('captureSnapshot retries pending tap outcome before post-gesture stabilization', async () => {
-  const sessionName = 'android-maestro-tap-outcome-before-stabilization';
-  const session = makeSession(sessionName, androidDevice);
-  const baselineNodes = [
-    {
-      ref: 'e1',
-      index: 0,
-      depth: 0,
-      type: 'android.widget.Button',
-      label: 'Navigate to Third',
-      hittable: true,
-      rect: { x: 302, y: 1301, width: 476, height: 110 },
-    },
-  ];
-  session.snapshot = {
-    nodes: baselineNodes,
-    createdAt: Date.now(),
-    backend: 'android',
-  };
-  session.pendingInteractionOutcome = {
-    action: 'click',
-    command: 'press',
-    positionals: ['540', '1356'],
-    flags: { platform: 'android' },
-    markedAt: Date.now(),
-    attemptsRemaining: 2,
-    preSignature: [
-      {
-        key: '|Navigate to Third||android.widget.Button||enabled|unselected|hittable|#0',
-        x: 302,
-        y: 1301,
-        width: 476,
-        height: 110,
-        discriminating: true,
-      },
-    ],
-  };
-  session.postGestureStabilization = {
-    action: 'click',
-    positionals: [],
-    markedAt: Date.now(),
-  };
-
-  let pressed = false;
-  mockTapPoint.mockImplementation(async () => {
-    pressed = true;
-    return { clicked: true };
-  });
-  legacyDispatchCapture.mockImplementation(async () => {
-    return {
-      nodes: !pressed
-        ? baselineNodes
-        : [
-            {
-              index: 0,
-              depth: 0,
-              type: 'android.widget.TextView',
-              label: 'Tab Third (3)',
-              rect: { x: 390, y: 884, width: 300, height: 55 },
-            },
-          ],
-      backend: 'android',
-    };
-  });
-
-  const result = await captureSnapshot({
-    device: androidDevice,
-    session,
-    flags: { snapshotInteractiveOnly: true },
-    logPath: '/tmp/daemon.log',
-    ...getRuntimeBindings(),
-  });
-
-  expect(result.snapshot.nodes).toEqual(
-    expect.arrayContaining([expect.objectContaining({ label: 'Tab Third (3)' })]),
-  );
-  // R58: the retry re-fires through the bound `tapPoint`, on the recorded coordinate pair.
-  expect(mockTapPoint).toHaveBeenCalledTimes(1);
-  expect(mockTapPoint.mock.calls[0]?.[0]?.point).toEqual({ x: 540, y: 1356 });
-  expect(session.pendingInteractionOutcome).toBeUndefined();
-  expect(session.postGestureStabilization).toBeUndefined();
-});
-
-test('captureSnapshot composes post-gesture stabilization with Android freshness capture', async () => {
-  const sessionName = 'android-post-gesture-freshness';
-  const baselineNodes = inboxBaselineNodes(18);
-  const changedNodes = buildNodes(
-    androidTextRows(18, (row) => (row === 1 ? 'album-0' : `Album row ${row}`)),
-  );
-  const session = makeAndroidFreshnessSession(sessionName, 'click', baselineNodes);
-  session.postGestureStabilization = {
-    action: 'click',
-    positionals: [],
-    markedAt: Date.now(),
-  };
-
-  legacyDispatchCapture
-    .mockResolvedValueOnce(androidCapture(baselineNodes, { rawNodeCount: 18, maxDepth: 1 }))
-    .mockResolvedValueOnce(androidCapture(changedNodes, { rawNodeCount: 18, maxDepth: 1 }))
-    .mockResolvedValueOnce(androidCapture(changedNodes, { rawNodeCount: 18, maxDepth: 1 }));
-
-  const result = await captureSnapshot({
-    device: androidDevice,
-    session,
-    flags: { snapshotInteractiveOnly: true },
-    logPath: '/tmp/daemon.log',
-  });
-
-  expect(result.snapshot.nodes).toEqual(
-    expect.arrayContaining([expect.objectContaining({ label: 'album-0' })]),
-  );
-  expect(legacyDispatchCapture.mock.calls.map((call) => call[1])).toEqual([
-    'snapshot',
-    'snapshot',
-    'snapshot',
-  ]);
-  expect(session.androidSnapshotFreshness).toBeUndefined();
-  expect(session.postGestureStabilization).toBeUndefined();
-});
-
-test('captureSnapshot composes pending outcome retry with Android freshness capture', async () => {
-  const sessionName = 'android-lazy-outcome-freshness';
-  const baselineNodes = inboxBaselineNodes(18);
-  const session = makeAndroidFreshnessSession(sessionName, 'click', baselineNodes);
-  session.pendingInteractionOutcome = {
-    action: 'click',
-    command: 'press',
-    positionals: ['180', '330'],
-    flags: { platform: 'android' },
-    markedAt: Date.now(),
-    attemptsRemaining: 2,
-    preSignature: buildInteractionSurfaceSignature(baselineNodes),
-  };
-
-  legacyDispatchCapture
-    .mockResolvedValueOnce(androidCapture([], { rawNodeCount: 18, maxDepth: 1 }))
-    .mockResolvedValueOnce(
-      androidCapture(
-        [
-          {
-            index: 0,
-            depth: 0,
-            type: 'android.widget.Button',
-            label: 'Create document',
-            hittable: true,
-          },
-        ],
-        { rawNodeCount: 1, maxDepth: 0 },
-      ),
-    );
-
-  const result = await captureSnapshot({
-    device: androidDevice,
-    session,
-    flags: { snapshotInteractiveOnly: true },
-    logPath: '/tmp/daemon.log',
-  });
-
-  expect(result.snapshot.nodes).toEqual(
-    expect.arrayContaining([expect.objectContaining({ label: 'Create document' })]),
-  );
-  expect(result.freshness).toEqual({
-    action: 'click',
-    retryCount: 1,
-    staleAfterRetries: false,
-    reason: undefined,
-  });
-  expect(legacyDispatchCapture.mock.calls.map((call) => call[1])).toEqual(['snapshot', 'snapshot']);
-  expect(session.pendingInteractionOutcome).toBeUndefined();
-  expect(session.androidSnapshotFreshness).toBeUndefined();
-});
-
-test('wait text on Android uses freshness-aware capture instead of one-shot snapshot polling', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'android-wait-freshness';
-  const session = makeAndroidFreshnessSession(sessionName, 'press', inboxBaselineNodes(18));
-  sessionStore.set(sessionName, session);
-
-  legacyDispatchCapture
-    .mockResolvedValueOnce(
-      androidCapture(androidTextRows(18, inboxRow), { rawNodeCount: 18, maxDepth: 1 }),
-    )
-    .mockResolvedValueOnce(
-      androidCapture(
-        [
-          { index: 0, depth: 0, type: 'android.widget.TextView', label: 'Create document' },
-          { index: 1, depth: 0, type: 'android.widget.TextView', label: 'Done' },
-        ],
-        { rawNodeCount: 2, maxDepth: 1 },
-      ),
-    );
-
-  // The wait budget includes Android's 250 ms freshness retry delay.
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'wait', { positionals: ['Create document', '500'] }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-    ...snapshotRuntimeFixture(),
-  });
-
-  expect(response?.ok).toBe(true);
-  if (response?.ok) {
-    expect(response.data?.text).toBe('Create document');
-  }
-  expect(legacyDispatchCapture).toHaveBeenCalledTimes(2);
-  expect(sessionStore.get(sessionName)?.snapshot?.nodes).toEqual(
-    expect.arrayContaining([expect.objectContaining({ label: 'Create document' })]),
-  );
-});
-
-test('wait text timeout includes compact current-surface labels and buttons', async () => {
-  const sessionName = 'android-wait-timeout-surface';
-  legacyDispatchCapture.mockResolvedValue({
-    nodes: locationPermissionNodes,
-    truncated: false,
-    backend: 'android',
-    analysis: { rawNodeCount: 4, maxDepth: 1 },
-  });
-
-  const response = await runWaitCommand(sessionName, androidDevice, ['Receipt uploaded', '50']);
-
-  expect(response?.ok).toBe(false);
-  if (response && !response.ok) {
-    expect(response.error.message).toBe(
-      'wait timed out for text: Receipt uploaded. Current surface: Location permission, Allow location access?, Not now, Continue.',
-    );
-    expect(response.error.details?.currentSurface).toEqual({
-      labels: ['Location permission', 'Allow location access?', 'Not now', 'Continue'],
-      buttons: ['Not now', 'Continue'],
-    });
-  }
-});
-
-test('wait selector timeout includes compact current-surface details', async () => {
-  const sessionName = 'android-wait-selector-timeout-surface';
-  legacyDispatchCapture.mockResolvedValue(locationRequiredCapture());
-
-  const response = await runWaitCommand(sessionName, androidDevice, ['id=receipt-uploaded', '50']);
-
-  expect(response?.ok).toBe(false);
-  if (response && !response.ok) {
-    expect(response.error.message).toBe(
-      'wait timed out for selector: id=receipt-uploaded. Current surface: Location required, Dismiss.',
-    );
-    expect(response.error.details?.currentSurface).toEqual({
-      labels: ['Location required', 'Dismiss'],
-      buttons: ['Dismiss'],
-    });
-  }
-});
-
-test('wait selector polling skips hidden-content hint derivation on every poll (#1270)', async () => {
-  // The #1270 repro: `wait 'label="Battery"' 8000` on Android. A presence-only wait never
-  // consumes scroll hints, so every per-poll snapshot capture must disable hint derivation —
-  // otherwise a pathological `dumpsys activity top` call is charged against the wait budget.
-  const sessionName = 'android-wait-selector-skips-hints';
-  legacyDispatchCapture
-    .mockResolvedValueOnce(locationRequiredCapture())
-    .mockResolvedValueOnce(batteryCapture());
-
-  const response = await runWaitCommand(sessionName, androidDevice, ['label="Battery"', '8000']);
-
-  expect(response?.ok).toBe(true);
-  const snapshotCalls = legacyDispatchCapture.mock.calls.filter(
-    ([, command]) => command === 'snapshot',
-  );
-  expect(snapshotCalls.length).toBe(2);
-  for (const call of snapshotCalls) {
-    const context = call[4] as { snapshotIncludeHiddenContentHints?: boolean } | undefined;
-    expect(context?.snapshotIncludeHiddenContentHints).toBe(false);
-  }
-});
-
-test('wait text polling skips hidden-content hint derivation on every poll (#1270)', async () => {
-  const sessionName = 'android-wait-text-skips-hints';
-  legacyDispatchCapture
-    .mockResolvedValueOnce(locationRequiredCapture())
-    .mockResolvedValueOnce(batteryCapture());
-
-  const response = await runWaitCommand(sessionName, androidDevice, ['Battery', '8000']);
-
-  expect(response?.ok).toBe(true);
-  const snapshotCalls = legacyDispatchCapture.mock.calls.filter(
-    ([, command]) => command === 'snapshot',
-  );
-  expect(snapshotCalls.length).toBe(2);
-  for (const call of snapshotCalls) {
-    const context = call[4] as { snapshotIncludeHiddenContentHints?: boolean } | undefined;
-    expect(context?.snapshotIncludeHiddenContentHints).toBe(false);
-  }
-});
-
-test('wait timeout summary prefers content labels over chrome and identifier noise', async () => {
-  const sessionName = 'ios-wait-timeout-surface-summary';
-  mockRunnerCommand.mockResolvedValue({ found: false });
-  legacyDispatchCapture.mockResolvedValue({
-    nodes: iosSurfaceSummaryNodes,
-    truncated: false,
-    backend: 'xctest',
-  });
-
-  const response = await runWaitCommand(sessionName, iosSimulatorDevice, [
-    'Impossible success text',
-    '50',
-  ]);
-
-  expect(response?.ok).toBe(false);
-  if (response && !response.ok) {
-    expect(response.error.message).toBe(
-      'wait timed out for text: Impossible success text. Current surface: Confirm catalog refresh, Keep browsing.',
-    );
-    expect(response.error.details?.currentSurface).toEqual({
-      labels: [
-        'Confirm catalog refresh',
-        'Keep browsing',
-        'host.exp.exponent:id/reload_button',
-        'Expo Go',
-        'gearshape.fill',
-        'Tab Bar',
-      ],
-      buttons: ['Keep browsing', 'host.exp.exponent:id/reload_button'],
-    });
-  }
-});
-
-test('wait timeout without readable capture does not inspect the current surface', async () => {
-  const sessionName = 'android-wait-timeout-surface-fails';
-  legacyDispatchCapture.mockRejectedValue(new Error('snapshot unavailable'));
-
-  const response = await runWaitCommand(sessionName, androidDevice, ['Receipt uploaded', '0']);
-
-  expect(response?.ok).toBe(false);
-  if (response && !response.ok) {
-    expect(response.error.message).toBe('wait timed out for text: Receipt uploaded');
-    expect(response.error.details?.reason).toBe('wait_capture_stalled');
-    expect(response.error.details?.retriable).toBe(true);
-    expect(response.error.details?.readableCaptures).toBe(0);
-  }
-  expect(legacyDispatchCapture).not.toHaveBeenCalled();
-});
-
 test('diff rejects unsupported kind', async () => {
   const sessionStore = makeSessionStore();
   const response = await handleSnapshotCommands({
@@ -1471,65 +430,4 @@ test('diff screenshot is not handled daemon-side (client-backed command)', async
     expect(response.error.code).toBe('INVALID_ARGS');
     expect(response.error.message).toMatch(/diff.*supports.*snapshot/i);
   }
-});
-
-test('wait selector bypasses a fresh matching session snapshot', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'android-wait-fresh-capture';
-  const session = makeSession(sessionName, androidDevice);
-  session.snapshot = {
-    createdAt: Date.now(),
-    presentationKey: buildSnapshotPresentationKey({}),
-    nodes: [
-      {
-        ref: 'e1',
-        index: 0,
-        type: 'android.widget.TextView',
-        label: 'Ready',
-      },
-    ],
-  };
-  sessionStore.set(sessionName, session);
-  legacyDispatchCapture.mockResolvedValue({
-    nodes: [
-      {
-        index: 0,
-        type: 'android.widget.TextView',
-        label: 'Ready',
-      },
-    ],
-  });
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'wait', { positionals: ['label="Ready"', '5000'] }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-    ...snapshotRuntimeFixture(),
-  });
-
-  expect(response?.ok).toBe(true);
-  expect(legacyDispatchCapture).toHaveBeenCalledWith(
-    expect.anything(),
-    'snapshot',
-    [],
-    undefined,
-    expect.anything(),
-  );
-});
-
-test('wait sleep bypasses sessionless runner cleanup wrapper', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'ios-sim';
-  sessionStore.set(sessionName, makeSession(sessionName, iosSimulatorDevice));
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'wait', { positionals: ['0'] }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response).toBeTruthy();
-  expect(response?.ok).toBe(true);
 });
