+++ src/daemon/replay/internal/__tests__/session-replay-repair-transaction.test.ts
@@ -16,9 +16,16 @@
  * abandoned close.
  * Fix 3: the source plan's terminal `close` is skipped while repair-armed, so
  * the resume completes instead of diverging on lifecycle.
+ *
+ * This file covers the core arm/resume/reap lifecycle and a commit-failure
+ * retry. `--force`/`--overwrite` preflight (#1258) and platform-close
+ * ordering (BLOCKER 2/3 sequencing) live in the sibling
+ * `session-replay-repair-transaction-force.test.ts` and
+ * `session-replay-repair-transaction-close-ordering.test.ts` — both split
+ * off this file once it crossed the module-size tripwire
+ * (docs/agents/testing.md).
  */
 import { test, expect, vi, beforeEach } from 'vitest';
-import { mkdtempForTestSync } from '../../../../__tests__/test-utils/tmp-dir.ts';
 
 vi.mock('@agent-device/platform-apple/runner/operations', () => ({
   prewarmIosRunnerSession: vi.fn(),
@@ -48,77 +55,31 @@ vi.mock('../../../snapshot-interactor-capture.ts', () => ({
 import fs from 'node:fs';
 import path from 'node:path';
 import { runReplayForTest } from '../../__tests__/replay-command-fixture.ts';
-import { handleSessionCloseCommands as handleProductionCloseCommand } from '../../../session-lifecycle/index.ts';
-import { SessionStore } from '../../../session-store.ts';
-import { LeaseRegistry } from '../../../lease-registry.ts';
 import { captureSnapshotWithInteractor } from '../../../snapshot-interactor-capture.ts';
 import {
   legacyDispatchCapture,
   resetLegacySnapshotCapture,
 } from '../../../__tests__/legacy-snapshot-capture-fixture.ts';
 import { dispatchApplicationLifecycleEffect } from '../../../__tests__/application-lifecycle-runtime-fixture.ts';
-import { AppError } from '@agent-device/kernel/errors';
-import {
-  makeIosSession,
-  makeRepairCompleteSession,
-  repairPublication,
-} from '../../../../__tests__/test-utils/session-factories.ts';
 import { HEAL_COMPLETE_SENTINEL } from '../../../session-script-writer.ts';
-import {
-  repairSessionBoundary,
-  repairSessionSourcePath,
-  markRepairTransactionComplete,
-} from '../../../session-replay-transaction.ts';
-import {
-  NO_SCRIPT_PUBLICATION,
-  scriptTargetPath,
-  scriptTargetForce,
-} from '../../../session-script-publication-state.ts';
+import { repairSessionSourcePath } from '../../../session-replay-transaction.ts';
 import { parseReplayScriptDetailed } from '@agent-device/ad-script';
-import type { DaemonRequest } from '../../../daemon-request.ts';
-import type { SessionState } from '../../../session-state.ts';
 import {
   baseReplayRequest as baseReq,
   writeReplayFile,
 } from '../../__tests__/session-replay-runtime.fixtures.ts';
-import { freshEvidence, makeRecordingReplayInvoke } from './session-replay-repair.fixtures.ts';
 import {
-  bindLifecycleRuntime,
-  inspectLifecycleRuntimeFacts,
-} from '../../../__tests__/application-lifecycle-runtime-harness.ts';
-import { platformResourceCleanup } from '../../../../platform-runtime-resource-cleanup.ts';
+  freshEvidence,
+  handleCloseCommand,
+  makeCompleteRepairSession,
+  makeRecordingReplayInvoke,
+  SAVE_ANNOTATION,
+  setup,
+} from './session-replay-repair.fixtures.ts';
 
 const mockDispatchCommand = legacyDispatchCapture;
 const mockCaptureSnapshotWithInteractor = vi.mocked(captureSnapshotWithInteractor);
 const mockLifecycleDispatch = vi.mocked(dispatchApplicationLifecycleEffect);
-
-function handleCloseCommand(
-  params: Omit<Parameters<typeof handleProductionCloseCommand>[0], 'inspectFacts' | 'bindDevice'>,
-) {
-  return handleProductionCloseCommand({
-    ...params,
-    platformResourceCleanup,
-    inspectFacts: inspectLifecycleRuntimeFacts,
-    bindDevice: bindLifecycleRuntime,
-  });
-}
-
-/** The persisted per-target overwrite grant (#1258), or `false` outside any publication. */
-function sessionTargetForce(session: SessionState | undefined): boolean {
-  return scriptTargetForce(session?.scriptPublication ?? NO_SCRIPT_PUBLICATION);
-}
-
-/** The explicit publication target path, or `undefined` for none/default. */
-function sessionTargetPath(session: SessionState | undefined): string | undefined {
-  return scriptTargetPath(session?.scriptPublication ?? NO_SCRIPT_PUBLICATION);
-}
-
-/** The recorded platform-close receipt identity, or `undefined` outside a repair transaction. */
-function sessionCloseReceipt(session: SessionState | undefined): string | undefined {
-  return session?.scriptPublication?.kind === 'repair'
-    ? session.scriptPublication.closeReceipt
-    : undefined;
-}
 
 beforeEach(() => {
   resetLegacySnapshotCapture(mockCaptureSnapshotWithInteractor);
@@ -141,23 +102,6 @@ beforeEach(() => {
     backend: 'xctest',
   });
 });
-
-const SAVE_ANNOTATION =
-  '# agent-device:target-v1 {"id":"save","role":"button","label":"Save","ancestry":[],"sibling":0,"viewportOrder":0,"verification":"verified"}';
-
-function setup(prefix: string) {
-  const root = mkdtempForTestSync(prefix);
-  const sessionStore = new SessionStore(path.join(root, 'sessions'));
-  const sessionName = 'default';
-  sessionStore.set(sessionName, makeIosSession(sessionName, { appBundleId: 'com.example.app' }));
-  return {
-    root,
-    sessionStore,
-    sessionName,
-    logPath: path.join(root, 'daemon.log'),
-    leaseRegistry: new LeaseRegistry(),
-  };
-}
 
 test('end-to-end repair transaction: cold divergence stays alive, corrective resume completes, close --save-script finalizes a COMPLETE healed .ad atomically, and an abandoned repair leaves no partial file', async () => {
   // ============================================================
@@ -475,27 +419,6 @@ test('BLOCKER 1: a --from continuation on a reaped session returns SESSION_NOT_F
   expect(resumed.error.code).not.toBe('REPLAY_DIVERGENCE');
 });
 
-/** A COMPLETE, committable repair-armed session at the default healed sibling path. */
-function makeCompleteRepairSession(sessionStore: SessionStore, sessionName: string, root: string) {
-  const session = makeRepairCompleteSession(sessionName, {
-    appBundleId: 'com.example.app',
-    scriptPublication: repairPublication('complete', { path: path.join(root, 'flow.healed.ad') }),
-    actions: [
-      { ts: 1, command: 'open', positionals: ['Demo'], flags: {} },
-      {
-        ts: 2,
-        command: 'press',
-        positionals: ['@e7'],
-        flags: {},
-        result: { selectorChain: ['id="save-v2"'] },
-        targetEvidence: freshEvidence('save-v2', 'Save V2'),
-      },
-    ],
-  });
-  sessionStore.set(sessionName, session);
-  return session;
-}
-
 test('BLOCKER 2b/2c: a close whose commit FAILS (no-clobber) keeps the session for retry and surfaces a distinct error', async () => {
   const { root, sessionStore, sessionName, logPath, leaseRegistry } = setup(
     'agent-device-repair-transaction-commit-fail-',
@@ -559,644 +482,4 @@ test('BLOCKER 2b/2c: a close whose commit FAILS (no-clobber) keeps the session f
   if (retry.ok) expect(retry.data?.savedScript).toBe(retryPath);
   const promoted = parseReplayScriptDetailed(fs.readFileSync(retryPath, 'utf8'));
   expect(promoted.actions.filter((a) => a.command === 'close')).toHaveLength(1);
-});
-
-// --- #1258: `--force`/`--overwrite` on `--save-script`. ---
-
-test('#1258: close --save-script --force overwrites an existing COMPLETE healed .ad instead of refusing', async () => {
-  const { root, sessionStore, sessionName, logPath, leaseRegistry } = setup(
-    'agent-device-repair-transaction-force-overwrite-',
-  );
-  makeCompleteRepairSession(sessionStore, sessionName, root);
-  // A prior COMPLETE (sentinel-marked) healed artifact already sits at the
-  // default path — `--force` must overwrite it instead of refusing.
-  fs.writeFileSync(
-    path.join(root, 'flow.healed.ad'),
-    `context platform=ios device="x"\nclick id="old"\n${HEAL_COMPLETE_SENTINEL}\n`,
-  );
-
-  const closeResponse = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: [],
-      flags: { force: true },
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-
-  expect(closeResponse.ok).toBe(true);
-  expect(sessionStore.get(sessionName)).toBeUndefined();
-  const healedPath = path.join(root, 'flow.healed.ad');
-  if (closeResponse.ok) expect(closeResponse.data?.savedScript).toBe(healedPath);
-  const script = fs.readFileSync(healedPath, 'utf8');
-  expect(script).toContain(HEAL_COMPLETE_SENTINEL);
-  const parsed = parseReplayScriptDetailed(script);
-  // The new repair's own actions replaced the stale prior artifact — the old
-  // content is gone, never merged/appended.
-  expect(parsed.actions.some((a) => a.positionals[0] === 'id="old"')).toBe(false);
-  expect(parsed.actions.map((a) => a.command)).toEqual(['open', 'press', 'close']);
-});
-
-test('#1258 arm-time preflight: an existing --save-script target rejects BEFORE any step runs', async () => {
-  const { root, sessionStore, sessionName, logPath } = setup(
-    'agent-device-repair-transaction-arm-preflight-',
-  );
-  const filePath = writeReplayFile(root, ['open "Demo" --relaunch', 'close']);
-  // The default healed sibling already exists — a prior repair, or an
-  // unrelated stale file sitting at that path.
-  fs.writeFileSync(path.join(root, 'flow.healed.ad'), 'context platform=ios device="x"\n');
-
-  const spy: DaemonRequest[] = [];
-  const invoke = makeRecordingReplayInvoke({ sessionStore, sessionName, spy });
-
-  const result = await runReplayForTest({
-    req: baseReq({ positionals: [filePath], flags: { saveScript: true } }),
-    sessionName,
-    logPath,
-    sessionStore,
-    invoke,
-  });
-
-  expect(result.ok).toBe(false);
-  if (!result.ok) {
-    expect(result.error.code).toBe('COMMAND_FAILED');
-    expect(result.error.message).toMatch(/already exists/);
-  }
-  // Not a single step dispatched — the preflight fired BEFORE the loop, not
-  // merely at the end (publish time). Even `open` never ran.
-  expect(spy).toHaveLength(0);
-  // The session was never armed (no boundary stamped) either — the whole
-  // repair-arm side effect is skipped, not just the dispatch.
-  expect(repairSessionBoundary(sessionStore.get(sessionName))).toBeUndefined();
-});
-
-test('#1258: --force skips the arm-time preflight and the replay proceeds despite the existing target', async () => {
-  const { root, sessionStore, sessionName, logPath } = setup(
-    'agent-device-repair-transaction-arm-preflight-force-',
-  );
-  const filePath = writeReplayFile(root, ['open "Demo" --relaunch', 'close']);
-  fs.writeFileSync(path.join(root, 'flow.healed.ad'), 'context platform=ios device="x"\n');
-
-  const spy: DaemonRequest[] = [];
-  const invoke = makeRecordingReplayInvoke({ sessionStore, sessionName, spy });
-
-  const result = await runReplayForTest({
-    req: baseReq({ positionals: [filePath], flags: { saveScript: true, force: true } }),
-    sessionName,
-    logPath,
-    sessionStore,
-    invoke,
-  });
-
-  expect(result.ok).toBe(true);
-  // `open` dispatched — the preflight did not block it. The terminal `close`
-  // is skipped while repair-armed (existing, unrelated behavior), so `open`
-  // is the only step this minimal script actually dispatches.
-  expect(spy.map((r) => r.command)).toEqual(['open']);
-  // `force` is persisted on the session from arm time, so a LATER commit
-  // (e.g. a bare `close`, or teardown) still honors the overwrite.
-  expect(sessionTargetForce(sessionStore.get(sessionName))).toBe(true);
-});
-
-test('#1258 preflight honors PERSISTED force: a --from continuation without --force is NOT rejected on an existing target a prior --force leg authorized', async () => {
-  const { root, sessionStore, sessionName, logPath } = setup(
-    'agent-device-repair-transaction-preflight-persisted-force-',
-  );
-  const filePath = writeReplayFile(root, [
-    'open "Demo" --relaunch',
-    SAVE_ANNOTATION,
-    'click id="save"',
-    'click id="confirm"',
-    'close',
-  ]);
-  // The healed target already exists — the preflight WOULD reject a
-  // continuation that only saw the (absent, this leg) LIVE force flag.
-  fs.writeFileSync(path.join(root, 'flow.healed.ad'), 'context platform=ios device="x"\n');
-  const invoke = makeRecordingReplayInvoke({
-    sessionStore,
-    sessionName,
-    evidence: (req) => (req.command === 'click' ? freshEvidence('confirm', 'Confirm') : undefined),
-  });
-
-  // Leg 1: `replay --save-script --force` — the LIVE force passes the preflight
-  // (target exists), arms the session, PERSISTS saveScriptForce, then diverges
-  // (nothing published, so the pre-existing target survives).
-  const leg1 = await runReplayForTest({
-    req: baseReq({ positionals: [filePath], flags: { saveScript: true, force: true } }),
-    sessionName,
-    logPath,
-    sessionStore,
-    invoke,
-  });
-  expect(leg1.ok).toBe(false);
-  if (leg1.ok) return;
-  expect(leg1.error.code).toBe('REPLAY_DIVERGENCE');
-  const divergence = leg1.error.details?.divergence as { resume: { planDigest: string } };
-  expect(sessionTargetForce(sessionStore.get(sessionName))).toBe(true);
-
-  // The agent's corrective press (blessed @ref), recorded live.
-  const session = sessionStore.get(sessionName)!;
-  sessionStore.recordAction(session, {
-    command: 'press',
-    positionals: ['@e7'],
-    flags: {},
-    result: { selectorChain: ['id="save-v2"'] },
-    targetEvidence: freshEvidence('save-v2', 'Save V2'),
-  });
-
-  // Leg 2: `replay --from N --save-script` WITHOUT --force — the target still
-  // exists. The persisted saveScriptForce must make the arm-time preflight use
-  // the SAME effective decision publication uses, so this is NOT rejected.
-  const leg2 = await runReplayForTest({
-    req: baseReq({
-      positionals: [filePath],
-      flags: { saveScript: true, replayFrom: 3, replayPlanDigest: divergence.resume.planDigest },
-    }),
-    sessionName,
-    logPath,
-    sessionStore,
-    invoke,
-  });
-  // Not rejected by the arm-time preflight (no "already exists"): the
-  // transaction reached completion instead.
-  expect(leg2.ok).toBe(true);
-  expect(sessionStore.get(sessionName)?.scriptPublication).toMatchObject({
-    kind: 'repair',
-    status: 'complete',
-  });
-  // A bare-boolean continuation never retargets, so force stays persisted.
-  expect(sessionTargetForce(sessionStore.get(sessionName))).toBe(true);
-});
-
-test('#1258 preflight is per-target: a --from continuation RETARGETING to an existing <b> WITHOUT live force is refused BEFORE dispatch, preserves the prior COMPLETE transaction, and a later close still commits the ORIGINAL <a>', async () => {
-  const { root, sessionStore, sessionName, logPath, leaseRegistry } = setup(
-    'agent-device-repair-transaction-preflight-retarget-',
-  );
-  const filePath = writeReplayFile(root, [
-    'open "Demo" --relaunch',
-    SAVE_ANNOTATION,
-    'click id="save"',
-    'click id="confirm"',
-    'close',
-  ]);
-  const targetA = path.join(root, 'a.ad');
-  const targetB = path.join(root, 'b.ad');
-  // <b> already exists (nobody opted to overwrite it); <a> does not, so leg 1
-  // arms/forces <a> cleanly.
-  fs.writeFileSync(targetB, 'context platform=ios device="x"\nclick id="unrelated"\n');
-  const beforeB = fs.readFileSync(targetB, 'utf8');
-  const spy: DaemonRequest[] = [];
-  const invoke = makeRecordingReplayInvoke({
-    sessionStore,
-    sessionName,
-    spy,
-    evidence: (req) => (req.command === 'click' ? freshEvidence('confirm', 'Confirm') : undefined),
-  });
-
-  // Leg 1: `--save-script=<a> --force` — arms + PERSISTS force for <a>, then
-  // diverges (nothing published).
-  const leg1 = await runReplayForTest({
-    req: baseReq({ positionals: [filePath], flags: { saveScript: targetA, force: true } }),
-    sessionName,
-    logPath,
-    sessionStore,
-    invoke,
-  });
-  expect(leg1.ok).toBe(false);
-  if (leg1.ok) return;
-  const divergence = leg1.error.details?.divergence as { resume: { planDigest: string } };
-  const session = sessionStore.get(sessionName)!;
-  expect(sessionTargetPath(session)).toBe(targetA);
-  expect(sessionTargetForce(session)).toBe(true);
-  // The agent's corrective press.
-  sessionStore.recordAction(session, {
-    command: 'press',
-    positionals: ['@e7'],
-    flags: {},
-    result: { selectorChain: ['id="save-v2"'] },
-    targetEvidence: freshEvidence('save-v2', 'Save V2'),
-  });
-  // The transaction for <a> has now reached COMPLETE (a resume leg ran to the
-  // end) — set directly here, mirroring `makeCompleteRepairSession`'s
-  // convention, to isolate THIS test's concern: a later retarget REJECTION must
-  // not corrupt this flag (BLOCKER 2 — the C2 `saveScriptComplete = false`
-  // reset must run AFTER the preflight's early-return, never before it).
-  markRepairTransactionComplete(session);
-  const dispatchesBeforeLeg2 = spy.length;
-
-  // Leg 2: `--from N --save-script=<b>` (explicit RETARGET, NO live force) — <b>
-  // exists. The persisted force was granted for <a>, and
-  // `applySaveScriptRetarget` WOULD clear it for <b>; the preflight must MATCH
-  // that per-target contract and REFUSE here, before any step dispatches,
-  // instead of executing the leg and only refusing at publish time.
-  const leg2 = await runReplayForTest({
-    req: baseReq({
-      positionals: [filePath],
-      flags: { saveScript: targetB, replayFrom: 3, replayPlanDigest: divergence.resume.planDigest },
-    }),
-    sessionName,
-    logPath,
-    sessionStore,
-    invoke,
-  });
-  expect(leg2.ok).toBe(false);
-  if (!leg2.ok) {
-    expect(leg2.error.code).toBe('COMMAND_FAILED');
-    expect(leg2.error.message).toMatch(/already exists/);
-  }
-  // Not a single step of leg 2 dispatched — refused BEFORE the loop.
-  expect(spy.length).toBe(dispatchesBeforeLeg2);
-  // READ-ONLY: the rejected request left the session target untouched — still
-  // armed/forced for <a>, never retargeted to <b>.
-  expect(sessionTargetPath(sessionStore.get(sessionName))).toBe(targetA);
-  expect(sessionTargetForce(sessionStore.get(sessionName))).toBe(true);
-  // BLOCKER 2 (a): the prior COMPLETE transaction SURVIVES the rejection — the
-  // C2 completion reset never ran, because the preflight returned first.
-  expect(sessionStore.get(sessionName)?.scriptPublication).toMatchObject({
-    kind: 'repair',
-    status: 'complete',
-  });
-  // <b> is byte-for-byte untouched.
-  expect(fs.readFileSync(targetB, 'utf8')).toBe(beforeB);
-
-  // BLOCKER 2 (b): a later bare `close` still COMMITS the ORIGINAL target <a>
-  // (never the rejected <b>) — proof the rejected retarget corrupted neither
-  // the completion flag nor the target path.
-  const closeResponse = await handleCloseCommand({
-    req: { token: 't', session: sessionName, command: 'close', positionals: [], flags: {} },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-  expect(closeResponse.ok).toBe(true);
-  if (closeResponse.ok) expect(closeResponse.data?.savedScript).toBe(targetA);
-  expect(fs.existsSync(targetA)).toBe(true);
-  const committed = parseReplayScriptDetailed(fs.readFileSync(targetA, 'utf8'));
-  expect(committed.actions.map((a) => a.command)).toEqual(['open', 'press', 'close']);
-  // <b> stayed untouched through the commit too.
-  expect(fs.readFileSync(targetB, 'utf8')).toBe(beforeB);
-});
-
-test('#1258 force is per-target: re-arming --save-script=<b> WITHOUT --force drops force persisted for <a>, so <b> is NOT overwritten', async () => {
-  const { root, sessionStore, sessionName, logPath, leaseRegistry } = setup(
-    'agent-device-repair-transaction-retarget-clears-force-',
-  );
-  // Armed and forced for target <a> (flow.healed.ad).
-  const session = makeCompleteRepairSession(sessionStore, sessionName, root);
-  session.scriptPublication = repairPublication('complete', {
-    boundary: 0,
-    path: path.join(root, 'flow.healed.ad'),
-    force: true,
-  });
-  // A DIFFERENT, unrelated file already sits at the retarget destination <b>
-  // (flow.promoted.ad) — nobody opted to overwrite THIS one.
-  const promotedPath = path.join(root, 'flow.promoted.ad');
-  fs.writeFileSync(
-    promotedPath,
-    `context platform=ios device="x"\nclick id="unrelated"\n${HEAL_COMPLETE_SENTINEL}\n`,
-  );
-  const before = fs.readFileSync(promotedPath, 'utf8');
-
-  // `close --save-script=<b>` (NO --force): retargeting from <a> to <b>
-  // without a live opt-in drops the force that was granted for <a>.
-  const closeResponse = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: [],
-      flags: { saveScript: promotedPath },
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-
-  // Refused — the retarget cleared the sticky force, so <b> is protected by the
-  // default no-clobber exactly like a fresh target would be.
-  expect(closeResponse.ok).toBe(false);
-  if (!closeResponse.ok) expect(closeResponse.error.message).toMatch(/already exists/);
-  // <b> is untouched, and the session is kept for retry.
-  expect(fs.readFileSync(promotedPath, 'utf8')).toBe(before);
-  expect(sessionStore.get(sessionName)).toBeDefined();
-  expect(sessionTargetForce(sessionStore.get(sessionName))).toBe(false);
-});
-
-test('#1258 force per-target, contrast: re-arming --save-script=<b> WITH --force DOES overwrite <b>', async () => {
-  const { root, sessionStore, sessionName, logPath, leaseRegistry } = setup(
-    'agent-device-repair-transaction-retarget-force-overwrites-',
-  );
-  const session = makeCompleteRepairSession(sessionStore, sessionName, root);
-  session.scriptPublication = repairPublication('complete', {
-    boundary: 0,
-    path: path.join(root, 'flow.healed.ad'),
-    force: true,
-  });
-  const promotedPath = path.join(root, 'flow.promoted.ad');
-  fs.writeFileSync(
-    promotedPath,
-    `context platform=ios device="x"\nclick id="unrelated"\n${HEAL_COMPLETE_SENTINEL}\n`,
-  );
-
-  // `close --save-script=<b> --force`: the live opt-in re-grants force for the
-  // new target, overwriting it.
-  const closeResponse = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: [],
-      flags: { saveScript: promotedPath, force: true },
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-
-  expect(closeResponse.ok).toBe(true);
-  const parsed = parseReplayScriptDetailed(fs.readFileSync(promotedPath, 'utf8'));
-  expect(parsed.actions.some((a) => a.positionals[0] === 'id="unrelated"')).toBe(false);
-});
-
-test('BLOCKER 2 (new): a repair close whose PLATFORM close fails never commits a healed .ad claiming a successful close', async () => {
-  const { root, sessionStore, sessionName, logPath, leaseRegistry } = setup(
-    'agent-device-repair-transaction-platform-close-fail-',
-  );
-  const session = makeCompleteRepairSession(sessionStore, sessionName, root);
-  const healedPath = path.join(root, 'flow.healed.ad');
-
-  // A targeted close (an explicit positional app target) is what makes
-  // `dispatchTargetedPlatformClose` actually dispatch instead of no-op.
-  const platformCloseError = new AppError('DEVICE_UNAVAILABLE', 'platform close failed', {
-    reason: 'device_disconnected',
-    hint: 'Reconnect the device and retry close.',
-    // BLOCKER 2 (second follow-up): the underlying platform error's own
-    // diagnosticId/logPath must survive normalization, never be flattened away.
-    diagnosticId: 'diag-platform-close-1',
-    logPath: '/tmp/platform-close-1.log',
-  });
-  mockLifecycleDispatch.mockRejectedValueOnce(platformCloseError);
-
-  const closeResponse = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: ['com.example.app'],
-      flags: {},
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-
-  // The prior implementation committed (recorded a successful `close` +
-  // published the healed .ad) BEFORE the platform close ran at all, so a
-  // failing platform close still left a COMMITTED artifact on disk claiming
-  // success, contradicting the failed-close lifecycle contract. Fixed: the
-  // platform close runs first, so a failure here means NOTHING was committed.
-  expect(fs.existsSync(healedPath)).toBe(false);
-  expect(closeResponse.ok).toBe(false);
-  if (!closeResponse.ok) {
-    expect(closeResponse.error.code).toBe('DEVICE_UNAVAILABLE');
-    // BLOCKER 3 (original): the session was kept for retry — `retriable`
-    // must agree. BLOCKER 2 (second follow-up): at the TOP level, and the
-    // underlying platform error's diagnosticId/logPath/details are preserved
-    // rather than discarded.
-    expect(closeResponse.error.retriable).toBe(true);
-    expect(closeResponse.error.details?.retriable).toBeUndefined();
-    expect(closeResponse.error.diagnosticId).toBe('diag-platform-close-1');
-    expect(closeResponse.error.logPath).toBe('/tmp/platform-close-1.log');
-    expect(closeResponse.error.details?.reason).toBe('device_disconnected');
-  }
-  // BLOCKER 2b-style contract: the session stays addressable, untouched, so
-  // the agent can fix the cause (e.g. reconnect the device) and retry.
-  expect(sessionStore.get(sessionName)).toBe(session);
-  expect(session.actions.some((a) => a.command === 'close')).toBe(false);
-
-  // Retry once the platform close succeeds: commits cleanly.
-  mockLifecycleDispatch.mockResolvedValueOnce(undefined);
-  const retry = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: ['com.example.app'],
-      flags: {},
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-  expect(retry.ok).toBe(true);
-  expect(fs.existsSync(healedPath)).toBe(true);
-  const healedScript = fs.readFileSync(healedPath, 'utf8');
-  expect(healedScript).toContain(HEAL_COMPLETE_SENTINEL);
-  expect(
-    parseReplayScriptDetailed(healedScript).actions.filter((a) => a.command === 'close'),
-  ).toHaveLength(1);
-});
-
-test('BLOCKER 3 (second follow-up): a retry after a SUCCESSFUL platform close but a FAILED commit never re-dispatches the platform close', async () => {
-  const { root, sessionStore, sessionName, logPath, leaseRegistry } = setup(
-    'agent-device-repair-transaction-close-idempotent-',
-  );
-  makeCompleteRepairSession(sessionStore, sessionName, root);
-  const healedPath = path.join(root, 'flow.healed.ad');
-  // A prior COMPLETE (sentinel-marked) healed artifact already sits at the
-  // default path — the commit must refuse to clobber it, giving a
-  // deterministic commit FAILURE after a platform close that genuinely
-  // succeeds (the lifecycle seam's `beforeEach` default resolves). A
-  // targeted close (an explicit positional app target) is what makes
-  // `dispatchTargetedPlatformClose` actually dispatch instead of no-op.
-  fs.writeFileSync(
-    healedPath,
-    `context platform=ios device="x"\nclick id="old"\n${HEAL_COMPLETE_SENTINEL}\n`,
-  );
-  const before = fs.readFileSync(healedPath, 'utf8');
-
-  const closeResponse = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: ['com.example.app'],
-      flags: {},
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-
-  // The platform close genuinely ran and succeeded; the SUBSEQUENT commit
-  // failed (no-clobber) — the session is retained for retry.
-  expect(mockLifecycleDispatch).toHaveBeenCalledTimes(1);
-  expect(closeResponse.ok).toBe(false);
-  if (!closeResponse.ok) expect(closeResponse.error.message).toMatch(/already exists/);
-  expect(sessionStore.get(sessionName)).toBeDefined();
-  expect(fs.readFileSync(healedPath, 'utf8')).toBe(before);
-  expect(sessionCloseReceipt(sessionStore.get(sessionName))).toBeDefined();
-
-  // Retry with an explicit path: the ALREADY-SUCCEEDED platform close must
-  // NEVER be dispatched again — a non-idempotent backend could fail (or
-  // wedge recovery entirely) on a second close of an already-closed target.
-  const retryPath = path.join(root, 'flow.promoted.ad');
-  const retry = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: ['com.example.app'],
-      flags: { saveScript: retryPath },
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-
-  // Still exactly ONE dispatch total — the retry consumed the recorded
-  // success and went straight to the commit.
-  expect(mockLifecycleDispatch).toHaveBeenCalledTimes(1);
-  expect(retry.ok).toBe(true);
-  expect(fs.existsSync(retryPath)).toBe(true);
-  const promoted = fs.readFileSync(retryPath, 'utf8');
-  expect(promoted).toContain(HEAL_COMPLETE_SENTINEL);
-});
-
-test('BLOCKER 3: a competing second writer never overwrites a COMPLETE artifact and gets a clear no-clobber error', async () => {
-  const { root, sessionStore, sessionName } = setup('agent-device-repair-transaction-competing-');
-  const healedPath = path.join(root, 'flow.healed.ad');
-
-  // Writer 1 commits a complete artifact at the default healed path.
-  const first = makeCompleteRepairSession(sessionStore, `${sessionName}-1`, root);
-  const r1 = sessionStore.writeSessionLog(first);
-  expect(r1.written).toBe(true);
-  const committed = fs.readFileSync(healedPath, 'utf8');
-  expect(committed).toContain(HEAL_COMPLETE_SENTINEL);
-
-  // Writer 2 (a second repair on the same source → same default path) attempts
-  // to publish over it. The atomic create-exclusive publish must FAIL rather
-  // than overwrite the complete artifact.
-  const second = makeCompleteRepairSession(sessionStore, `${sessionName}-2`, root);
-  second.actions[1] = {
-    ts: 2,
-    command: 'press',
-    positionals: ['@e9'],
-    flags: {},
-    result: { selectorChain: ['id="different"'] },
-    targetEvidence: freshEvidence('different', 'Different'),
-  };
-  const r2 = sessionStore.writeSessionLog(second);
-  expect(r2.written).toBe(false);
-  expect(r2.written === false && r2.error?.message).toMatch(/already exists/);
-  // The first writer's complete artifact is byte-for-byte intact.
-  expect(fs.readFileSync(healedPath, 'utf8')).toBe(committed);
-});
-
-test('BLOCKER 3 (third follow-up): an untargeted close that performed NO platform operation never lets a targeted retry skip the platform close', async () => {
-  const { root, sessionStore, sessionName, logPath, leaseRegistry } = setup(
-    'agent-device-repair-transaction-close-identity-untargeted-',
-  );
-  makeCompleteRepairSession(sessionStore, sessionName, root);
-  const healedPath = path.join(root, 'flow.healed.ad');
-  // A prior COMPLETE artifact already sits at the default path so the commit
-  // deterministically fails and the session is retained for retry.
-  fs.writeFileSync(
-    healedPath,
-    `context platform=ios device="x"\nclick id="old"\n${HEAL_COMPLETE_SENTINEL}\n`,
-  );
-
-  // First attempt: UNTARGETED close — `shouldDispatchPlatformClose` is false
-  // (no positional target, not `web`), so the platform close never dispatches
-  // at all; the commit then fails (no-clobber).
-  const first = await handleCloseCommand({
-    req: { token: 't', session: sessionName, command: 'close', positionals: [], flags: {} },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-  expect(first.ok).toBe(false);
-  expect(mockLifecycleDispatch).not.toHaveBeenCalled();
-  expect(sessionStore.get(sessionName)).toBeDefined();
-
-  const retry = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: ['com.example.app'],
-      flags: { saveScript: path.join(root, 'flow.promoted.ad') },
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-  expect(mockLifecycleDispatch).toHaveBeenCalledTimes(1);
-  expect(mockLifecycleDispatch.mock.calls[0]?.[2]).toEqual(['com.example.app']);
-  expect(retry.ok).toBe(true);
-});
-
-test('BLOCKER 3 (third follow-up): a retry targeting a DIFFERENT app than the succeeded close never skips the platform close for the new target', async () => {
-  const { root, sessionStore, sessionName, logPath, leaseRegistry } = setup(
-    'agent-device-repair-transaction-close-identity-changed-target-',
-  );
-  makeCompleteRepairSession(sessionStore, sessionName, root);
-  const healedPath = path.join(root, 'flow.healed.ad');
-  fs.writeFileSync(
-    healedPath,
-    `context platform=ios device="x"\nclick id="old"\n${HEAL_COMPLETE_SENTINEL}\n`,
-  );
-
-  // First attempt targets app-a; the platform close succeeds, the commit fails.
-  const first = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: ['com.example.app-a'],
-      flags: {},
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-  expect(first.ok).toBe(false);
-  expect(mockLifecycleDispatch).toHaveBeenCalledTimes(1);
-  expect(sessionCloseReceipt(sessionStore.get(sessionName))).toBeDefined();
-
-  // Retry targets a DIFFERENT app (app-b) — a genuinely different platform
-  // operation. The prior session-wide marker would wrongly treat app-b as
-  // already closed just because SOME close succeeded. It must dispatch again,
-  // against app-b specifically.
-  const retry = await handleCloseCommand({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'close',
-      positionals: ['com.example.app-b'],
-      flags: { saveScript: path.join(root, 'flow.promoted.ad') },
-    },
-    sessionName,
-    logPath,
-    sessionStore,
-    leaseRegistry,
-  });
-  expect(mockLifecycleDispatch).toHaveBeenCalledTimes(2);
-  expect(mockLifecycleDispatch.mock.calls[1]?.[2]).toEqual(['com.example.app-b']);
-  expect(retry.ok).toBe(true);
 });
